Use base64DecodeAndZlibInflate
